+++ include/fractals.hpp
@@ -39,11 +39,3 @@ FractalSet mandelbrot() {
 
 }  // namespace fractals
 
-int main() {
-    auto m_set = fractals::mandelbrot();
-    std::cout << "{ ";
-    for (auto i : m_set) {
-        std::cout << i << " ";
-    }
-    std::cout << "}" << std::endl;
-}
